refactor: move flush from commands directory since it is reserved directory now
